+++ app/Http/Controllers/CarouselController.php
@@ -8,8 +8,16 @@ use App\Http\Requests\UpdateCarouselRequest;
 
 class CarouselController extends Controller
 {
+    protected $permissionsArray = [];
+
+    public function __construct()
+    {
+        $this->permissionsArray = getPermissionsArray();
+    }
     public function index()
     {
+        if (!in_array('View Page Content carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $carousels = Carousel::paginate(5);
         return view('carousels.index', compact('carousels'));
     }
@@ -17,11 +25,15 @@ class CarouselController extends Controller
 
     public function create()
     {
+        if (!in_array('create carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('carousels.create');
     }
 
     public function store(StoreCarouselRequest $request)
     {
+        if (!in_array('create carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['subheading', 'heading', 'button_text', 'button_link']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -35,16 +47,22 @@ class CarouselController extends Controller
 
     public function show(Carousel $carousel)
     {
+        if (!in_array('View Page Content carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('carousels.show', compact('carousel'));
     }
 
     public function edit(Carousel $carousel)
     {
+        if (!in_array('edit carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('carousels.edit', compact('carousel'));
     }
 
     public function update(UpdateCarouselRequest $request, Carousel $carousel)
     {
+        if (!in_array('edit carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['subheading', 'heading', 'button_text', 'button_link']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -61,6 +79,8 @@ class CarouselController extends Controller
 
     public function destroy(Carousel $carousel)
     {
+        if (!in_array('delete carousels', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         if ($carousel->image_url) {
             \Storage::disk('public')->delete(str_replace('/storage/', '', $carousel->image_url));
         }
