+++ app/Http/Controllers/NavigationController.php
@@ -7,11 +7,19 @@ use App\Http\Requests\UpdateNavigationRequest;
 
 class NavigationController extends Controller
 {
+    protected $permissionsArray = [];
+
+    public function __construct()
+    {
+        $this->permissionsArray = getPermissionsArray();
+    }
     /**
      * Display a listing of the navigation items.
      */
     public function index()
     {
+        if (!in_array('View Page Content navigations', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $navigations = Navigation::paginate(5);
         return view('navigations.index', compact('navigations'));
     }
@@ -21,6 +29,8 @@ class NavigationController extends Controller
      */
     public function create()
     {
+        if (!in_array('add navigation', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('navigations.create');
     }
 
@@ -29,6 +39,8 @@ class NavigationController extends Controller
      */
     public function store(StoreNavigationRequest $request)
     {
+        if (!in_array('add navigation', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['label', 'url', 'order']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -41,6 +53,8 @@ class NavigationController extends Controller
      */
     public function show(Navigation $navigation)
     {
+        if (!in_array('View Page Content navigations', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('navigations.show', compact('navigation'));
     }
 
@@ -49,6 +63,8 @@ class NavigationController extends Controller
      */
     public function edit(Navigation $navigation)
     {
+        if (!in_array('edit navigation', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('navigations.edit', compact('navigation'));
     }
 
@@ -57,6 +73,8 @@ class NavigationController extends Controller
      */
     public function update(UpdateNavigationRequest $request, Navigation $navigation)
     {
+        if (!in_array('edit navigation', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['label', 'url', 'order']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -69,6 +87,8 @@ class NavigationController extends Controller
      */
     public function destroy(Navigation $navigation)
     {
+        if (!in_array('delete navigation', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $navigation->delete();
         return redirect()->route('navigations.index')->with('success', 'Navigation item deleted successfully!');
     }
