+++ app/Http/Controllers/RoleController.php
@@ -9,20 +9,32 @@ use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
+    protected $permissionsArray = [];
+
+    public function __construct()
+    {
+        $this->permissionsArray = getPermissionsArray();
+    }
     public function index()
     {
+        if (!in_array('View Roles', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $roles = Role::paginate(5);
         return view('roles.index', compact('roles'));
     }
 
     public function create()
     {
+        if (!in_array('create role', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $permissions = Permission::all();
         return view('roles.create', compact('permissions'));
     }
 
     public function store(StoreRoleRequest $request)
     {
+        if (!in_array('create role', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $role = Role::create($request->only('name'));
 
         if ($request->has('permissions')) {
@@ -34,11 +46,15 @@ class RoleController extends Controller
 
     public function show(Role $role)
     {
+        if (!in_array('View Roles', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('roles.show', compact('role'));
     }
 
     public function edit(Role $role)
     {
+        if (!in_array('edit role', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $permissions = Permission::all();
         $selectedPermissions = $role->permissions->pluck('id')->toArray();
         return view('roles.edit', compact('role', 'permissions', 'selectedPermissions'));
@@ -46,6 +62,8 @@ class RoleController extends Controller
 
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        if (!in_array('edit role', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $role->update($request->validated());
 
         if ($request->has('permissions')) {
@@ -57,6 +75,8 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        if (!in_array('delete role', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $role->delete();
         return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
     }
