+++ app/Http/Controllers/ServiceController.php
@@ -8,8 +8,16 @@ use App\Http\Requests\UpdateServiceRequest;
 
 class ServiceController extends Controller
 {
+    protected $permissionsArray = [];
+
+    public function __construct()
+    {
+        $this->permissionsArray = getPermissionsArray();
+    }
     public function index()
     {
+        if (!in_array('View Page Content services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $services = Service::paginate(5);
         return view('services.index', compact('services'));
     }
@@ -17,11 +25,15 @@ class ServiceController extends Controller
 
     public function create()
     {
+        if (!in_array('create services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('services.create');
     }
 
     public function store(StoreServiceRequest $request)
     {
+        if (!in_array('create services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['title', 'description']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -34,16 +46,22 @@ class ServiceController extends Controller
 
     public function show(Service $service)
     {
+        if (!in_array('View Page Content services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('services.show', compact('service'));
     }
 
     public function edit(Service $service)
     {
+        if (!in_array('edit services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('services.edit', compact('service'));
     }
 
     public function update(UpdateServiceRequest $request, Service $service)
     {
+        if (!in_array('edit services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['title', 'description']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -60,6 +78,8 @@ class ServiceController extends Controller
 
     public function destroy(Service $service)
     {
+        if (!in_array('delete services', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         if ($service->icon) {
             \Storage::disk('public')->delete($service->icon);
         }
