+++ app/Http/Controllers/TimelineEventController.php
@@ -8,19 +8,31 @@ use App\Http\Requests\UpdateTimelineEventRequest;
 
 class TimelineEventController extends Controller
 {
+    protected $permissionsArray = [];
+
+    public function __construct()
+    {
+        $this->permissionsArray = getPermissionsArray();
+    }
     public function index()
     {
+        if (!in_array('View Page Content timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $timelineEvents = TimelineEvent::paginate(5);
         return view('timeline-events.index', compact('timelineEvents'));
     }
 
     public function create()
     {
+        if (!in_array('create timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('timeline-events.create');
     }
 
     public function store(StoreTimelineEventRequest $request)
     {
+        if (!in_array('create timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['title', 'subheading', 'start_date', 'end_date']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -36,16 +48,22 @@ class TimelineEventController extends Controller
 
     public function show(TimelineEvent $timelineEvent)
     {
+        if (!in_array('View Page Content timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('timeline-events.show', compact('timelineEvent'));
     }
 
     public function edit(TimelineEvent $timelineEvent)
     {
+        if (!in_array('edit timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         return view('timeline-events.edit', compact('timelineEvent'));
     }
 
     public function update(UpdateTimelineEventRequest $request, TimelineEvent $timelineEvent)
     {
+        if (!in_array('edit timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         $data = $request->only(['title', 'subheading', 'start_date', 'end_date']);
         $data['is_active'] = $request->has('is_active') ? true : false;
 
@@ -64,6 +82,8 @@ class TimelineEventController extends Controller
 
     public function destroy(TimelineEvent $timelineEvent)
     {
+        if (!in_array('delete timeline', array_keys(getPermissionsArray())))
+            abort(403, 'You are not authorized to view this resource.');
         if ($timelineEvent->image) {
             \Storage::disk('public')->delete(str_replace('/storage/', '', $timelineEvent->image));
         }
